move all unit test near the file they test
