+++ lib/queries/helpers/batchParameter.spec.js
@@ -1,4 +1,6 @@
-import batchParameter from '../../lib/queries/batchParameter';
+import { assert } from 'chai';
+
+import batchParameter from './batchParameter';
 
 describe('batchParameter', () => {
     it('should return batchParameter', () => {
