+++ lib/queries/helpers/namedToNumericParameter.spec.js
@@ -1,4 +1,6 @@
-import namedToNumericParameter from '../../lib/queries/namedToNumericParameter';
+import { assert } from 'chai';
+
+import namedToNumericParameter from './namedToNumericParameter';
 
 describe('namedToNumericParameter', () => {
     it('should replace named parameters by number one', () => {
