+++ lib/queries/helpers/sanitizeParameter.spec.js
@@ -1,4 +1,5 @@
-import sanitizeParameter from '../../lib/queries/sanitizeParameter';
+import { assert } from 'chai';
+import sanitizeParameter from './sanitizeParameter';
 
 describe('sanitizeParameter', () => {
     it('should keep only attributes passed in first argument', () => {
