+++ lib/queries/helpers/valueSubQuery.spec.js
@@ -1,4 +1,5 @@
-import valueSubQuery from '../../lib/queries/valueSubQuery';
+import { assert } from 'chai';
+import valueSubQuery from './valueSubQuery';
 
 describe('valueSubQuery', () => {
     it('should return subQuery filtering out unwanted field', () => {
