+++ lib/queries/helpers/whereQuery.spec.js
@@ -1,4 +1,5 @@
-import whereQuery, * as whereQueryGet from '../../lib/queries/whereQuery';
+import { assert } from 'chai';
+import whereQuery, * as whereQueryGet from './whereQuery';
 
 describe('whereQuery', () => {
     it('should return whereQuery', () => {
