+++ lib/services/pg-pool.js
@@ -1,5 +1,5 @@
 import pg from 'pg';
-import namedToNumericParameter from '../queries/namedToNumericParameter';
+import namedToNumericParameter from '../queries/helpers/namedToNumericParameter';
 import arrayToLitteral from '../utils/arrayToLitteral';
 
 const iterate = async (query, it, data = it.next()) => {
