+++ lib/utils/arrayToLitteral.spec.js
