+++ lib/utils/combineLiterals.spec.js
