+++ lib/utils/middleware.spec.js
